root uninstall

+++ dnsinterceptor/uninstall.sh
@@ -1,4 +1,10 @@
 
+# Check if root user
+if [ `id -u` -ne 0 ]; then
+ echo "Uninstall as root user"
+ exit 1;
+fi
+
 # Stop/disable the service
 systemctl stop dnsinterceptor
 systemctl disable dnsinterceptor
